requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
pika==1.3.2
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
